requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
twilio
google-api-python-client
google-auth
pydantic
pydantic-settings
python-multipart
